Bulkupload

+++ src/main/java/com/yash/ytms/repository/NominationUploadRepository.java
@@ -6,6 +6,6 @@ import com.yash.ytms.domain.Event;
 import com.yash.ytms.domain.Nomination;
 import com.yash.ytms.dto.NominationData;
 
-public interface NominationUploadRepository extends JpaRepository<Nomination,Long>  {
+public interface NominationUploadRepository extends JpaRepository<Nomination,Integer>  {
 
 }
